tweak the url path logic to squash dup '/' in the path only
This fixes a weird edge case issue where you add a url value or
other value as a query string value and it gets its double forward
slashes squashed unintentionally.  The goal was to prevent
generating invalid urls but the multi forward slashes is only
nonsensical in the path info part of the url.  This switches the
logic to squash for the path info part but not for the query
string part.