add view handler `set_cookie` private helper
This is a helper method for setting cookie headers.  It uses
`Rack::Utils.set_cookie_header!` under the covers but its api is
different.  It takes a cookie name, cookie value, and cookie
options.  It passes its own `self.headers` to the rack utils
method for making sure the headers are updated.

We chose this API as it is more defined and explicit compared to
the rack utils method (so it will be easier to support and
document).  It also makes more intuitive sense as a cookie has
a name and value.  The options allow you to pass the same options
you would to the rack utils method to control and secure the
cookies.  This just provides a defined api and abstraction for
the rack utils implementation so we can change it as needed with
minimal app impact.