improved default response body handling
This is a set of updates to how the response body is defaulted and
handled when called with "empty" values.  The goal of all of this
is that the default response body will be used unless a meaningful
alternative response body has been given.

First off, this updates the default body to, formally, be an array
with a single empty string as this more accurately conforms to
Rack's spec that body objects must yield string values.

Next, this updates the `body` helper to ignore `nil` and empty
string values as they have no value and are identical to the
default body in intent.

Finally, this updates the halt method to also ignore `nil` and
empty string values for similar reasons as above.