add a handler DSL api for setting default response values
This allows individual handlers to override the default response
values for status, headers, and body in a reliable manner.
Specifically, this will be used in Deas-Json to override the
default body for its needs.

First this moves the status/headers/body defaults to the view
handler.  This also adds the DSL singleton methods for overriding
the defaults on a per handler basis.

Second, this reworks the runner logic to look to the given handler
class for the default values.

Note: to make this happen, I had to do two other things:

* I had to commonize the logic for setting body response values.
  This logic applies both for setting handler instance response
  bodies and handler-specific default response bodies.
* I had to rework some of the test contexts to work with the new
  singleton handler class settings.  This mostly meant reworking
  to not use a constant handler class and instead use an anonymous
  handler class so that any settings made in the tests are thrown
  away in subsequent tests.