trailing slashes: treat root index path `/` as a leading slash
This is an edge case I forgot to account for when originally
implementing the trailing slashes handling logic feature (see
PRs 239, 240, and 241).  The root index path, `/` should be
considered a *leading* slash not a trailing slash.

This updates both the router validation logic and the middleware
handling logic to account for this special case.  I also expanded
the middleware system tests and the router unit tests to explicitly
test for this special case.

In addition, this does a few cleanups I noticed while implementing
this:

* I chose to rework the variable names and boolean handling in the
  router validation logic to be more "positive boolean" focused.
  I think this makes it easier to read and is a convention to use
  positive boolean logic whenever possible.
* I renamed the middleware's `RequireNoHandler` to `RemoveHandler`.
  This should have been done in d389f98 but was missed.